pyproject.toml
[project]
name = "mcp-doc-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.0.0",
]

[tool.mcp]
entrypoint = "mcp_docs_agent.server:mcp"
